pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
openai = "^1.28.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
web-scraper = "^1.0"
flask = "^3.0.3"
embedchain = "^0.1.113"
python-dotenv = "^1.0.1"
spacy = "^3.7.5"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"